+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Head
+* Added: better handling of vendor prefixes in `property-unit-blacklist` and `property-unit-whitelist`, e.g. if you enter `animation` it now also checks `-webkit-animation`.
+* Added: support for regular expression property identification in `property-blacklist`, `property-unit-blacklist`, `property-unit-whitelist`, `property-value-blacklist`, and `property-whitelist`.
 * Added: brought `stylelint-rule-tester` into this repo, and exposed it at `stylelint.utils.ruleTester`.
 * Added: `codeFilename` option to Node API.
 * Added: `font-weight-notation` rule.
@@ -6,6 +8,7 @@
 * Added: `no-duplicate-selectors` rule.
 * Added: `no-unknown-animations` rule.
 * Added: `property-value-blacklist` rule.
+* Added: `property-value-whitelist` rule.
 * Added: `time-no-imperceptible` rule.
 * Added: `ignore: "descendant"` and `ignore: "compounded"` options for `selector-no-type`.
 * Added: support for using names of modules for the CLI's `--config` argument, not just paths.